Move `Eras` to `Cardano.Read.Ledger.Tx.Eras`
