+++ lib/read/lib/Cardano/Read/Ledger/Tx/Certificates.hs
@@ -36,6 +36,9 @@ import Cardano.Ledger.Crypto
 import Cardano.Ledger.Shelley.TxCert
     ( ShelleyTxCert
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -49,9 +52,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( view
