+++ lib/read/lib/Cardano/Read/Ledger/Tx/Outputs.hs
@@ -34,6 +34,9 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Shelley.TxOut
     ( ShelleyTxOut
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -47,9 +50,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( view
