+++ lib/read/lib/Cardano/Read/Ledger/Tx/TxId.hs
@@ -27,6 +27,9 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto
     ( StandardCrypto
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read
     ( Tx
     )
@@ -40,9 +43,6 @@ import Cardano.Wallet.Read.Eras
     , IsEra (..)
     , Mary
     , Shelley
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
