+++ lib/read/lib/Cardano/Wallet/Read/Tx/ExtraSigs.hs
@@ -31,6 +31,9 @@ import Cardano.Ledger.Keys
     ( KeyHash
     , KeyRole (..)
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -44,9 +47,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
