+++ lib/read/lib/Cardano/Wallet/Read/Tx/Fee.hs
@@ -30,6 +30,9 @@ import Cardano.Ledger.Core
     ( bodyTxL
     , feeTxBodyL
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -43,9 +46,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
