+++ lib/read/lib/Cardano/Wallet/Read/Tx/Hash.hs
@@ -36,15 +36,15 @@ import Cardano.Ledger.Core
 import Cardano.Ledger.TxIn
     ( TxId (..)
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read
     ( Tx
     )
 import Cardano.Wallet.Read.Eras
     ( Era (..)
     , IsEra (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
