+++ lib/read/lib/Cardano/Wallet/Read/Tx/ScriptValidity.hs
@@ -26,6 +26,9 @@ import Cardano.Ledger.Alonzo.Tx
     ( IsValid
     , isValidTxL
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -39,9 +42,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
