+++ lib/read/lib/Cardano/Wallet/Read/Tx/Validity.hs
@@ -38,6 +38,9 @@ import Cardano.Ledger.Shelley.TxBody
 import Cardano.Ledger.Slot
     ( SlotNo
     )
+import Cardano.Read.Ledger.Tx.Eras
+    ( onTx
+    )
 import Cardano.Wallet.Read.Eras
     ( Allegra
     , Alonzo
@@ -51,9 +54,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Tx
     ( Tx (..)
-    )
-import Cardano.Wallet.Read.Tx.Eras
-    ( onTx
     )
 import Control.Lens
     ( (^.)
